backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database - Supabase
supabase==2.3.4
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Services
openai==1.6.1
httpx==0.25.2

# Whisper AI - Full Production Support
faster-whisper==0.10.0
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1

# File handling
aiofiles==23.2.1

# WebSocket
websockets==12.0

# Additional production dependencies
requests==2.31.0
typing-extensions==4.8.0